move DomainController::localeTabsAction() from project-base to the framework package
