+++ packages/framework/src/Resources/views/Admin/Inline/Domain/locales.html.twig
